Changing IRLS for RVC
There is no range error, when running main. However, the solution
doesn't appear to be sparse. Furthermoer, using scipy.optimize results
in the same overflow in the same place. Namely in _likelihood.

We also need a way to get the class and class probability of the output,
since right now the output in main doesn't give us much.