Bugfix: Fixed pruning and prediction for regression
Pruning was not working correctly for the regression case. This is fixed now.
Prediction was not working either and did not catch the case of a pruned
bias weight.

Note: Accuracy and sparsity is hightly dependent on the initial values of
alpha. Please keep that in mind.
Open question: Is it intended to be that sensitive to the initial values of alpha?